# api.yaml
openapi: "3.0.3"
info:
  title: Sanshain API
  description: |
    API contract for tooling that talks to Sanshain (build plugins, CLIs).
    A **Producer** provides its complete specification under a producer-declared
    **version**; a **Consumer** pins an exact version and requires only the
    endpoint snippets it uses. Covers providing specifications (OpenAPI,
    AsyncAPI, gRPC/Proto) and requiring endpoint snippets. See CONTEXT.md for
    the vocabulary and docs/adr/0003-versions-replace-branches.md for the model.

    **The version lives in the spec file.** For OpenAPI and AsyncAPI it is read
    from `info.version`; for proto from a mandatory
    `// sanshain-version: MAJOR.MINOR.PATCH` comment. It must be strict
    three-part semver — no `v` prefix, no `-SNAPSHOT` or other suffixes.

    **Stability** is declared on every Provide: `snapshot` (overwritable
    work-in-progress, expires when unused) or `ga` (immutable; the number is
    permanently claimed). Resolution prefers GA, falls back only to the
    same-numbered snapshot, and otherwise fails immediately — no fallback, no
    long-polling.

    **2.0 is a clean break** of the 1.x branch-based contract: `branch`,
    `base_version`, `force`, `timeout`, `source_protected_branch` and
    `pull_from_branch` are no longer accepted anywhere and are rejected as
    unknown fields.

    **Legacy names**: callers may send `servicename` in place of `producername`
...
    the same value is rejected as a duplicate field.
...
servers:
...
paths:
...
  /producers/{producername}/versions:
    get:
      operationId: listProducerVersions
      summary: List a Producer's version lines
      description: |
        Every version-line entry of one Producer — version, stability, content
        hash, timestamps, last provider, endpoint count, and (for snapshots,
        when cleanup is enabled) the use-based expiry. What "what can I
        upgrade to?" tooling reads.
      parameters:
        - name: producername
          in: path
          required: true
          schema: { type: string }
        - name: api_type
          in: query
          required: false
          schema:
            type: string
            enum: [openapi, asyncapi, proto]
          description: Narrow to one API type's line.
      responses:
        "200":
          description: The version-line entries, API types grouped, newest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProducerVersionInfo"
        "404":
          description: Unknown Producer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
...
components:
  schemas:
    ErrorBody:
      type: object
      description: |
        The body every error response carries. `error` holds the human-readable
        message; `proposed_version` appears only on 409s from the version
        rules.
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable description of what went wrong.
        proposed_version:
          type: string
          description: |
            The next free version the Producer should publish as instead.
            Present only on a 409 refusing a Provide by the version rules.
    Stability:
      type: string
      enum: [snapshot, ga]
      description: |
        Which of exactly two states a stored version is in. Declared by the
        caller on every Provide; a state of the stored version, never part of
        the version string. Snapshots are overwritable and expire when unused;
        GA is immutable and permanently claims its number.
...
    ProducerVersionInfo:
      type: object
      required:
        - api_type
        - version
        - stability
        - content_hash
        - provided_by
        - created_at
        - updated_at
        - endpoint_count
      properties:
        api_type:
          type: string
          enum: [openapi, asyncapi, proto]
        version:
          type: string
        stability:
          $ref: "#/components/schemas/Stability"
        content_hash:
          type: string
        author:
          type: string
          nullable: true
          description: Unverified blame attribution recorded with the Provide.
        provided_by:
          type: string
          description: The authenticated Actor behind the last successful Provide.
        created_at:
          type: string
          description: When this version first appeared on the line.
        updated_at:
          type: string
          description: Last successful Provide (overwrite or promotion).
        last_required_at:
          type: string
          nullable: true
          description: Last successful require against this version.
        endpoint_count:
          type: integer
          format: int64
        expires_at:
          type: string
          nullable: true
          description: |
            Use-based expiry for snapshots when cleanup is enabled; absent for
            GA (never age-culled) or when cleanup is off.